pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbt-governance"
version = "0.1.0"
description = "Deterministic governance enforcement and AI-powered code review for dbt Cloud projects"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "dbt Governance Contributors" },
]
keywords = ["dbt", "governance", "data-engineering", "analytics-engineering", "best-practices", "ci", "linting", "legacy-migration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "sqlglot>=20.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
ai = ["anthropic>=0.39.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
    "ruff>=0.8.0",
]
all = ["dbt-governance[ai,dev]"]

[project.scripts]
dbt-governance = "dbt_governance.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/dbt-governance"
Documentation = "https://github.com/your-org/dbt-governance#readme"
Repository = "https://github.com/your-org/dbt-governance"
"Bug Tracker" = "https://github.com/your-org/dbt-governance/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dbt_governance"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
